ai-security-platform/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.4
aioredis==2.0.1
scikit-learn==1.4.2
numpy==1.26.4
sentence-transformers==2.7.0
httpx==0.27.0
slowapi==0.1.9
python-dotenv==1.0.1
structlog==24.1.0
prometheus-client==0.20.0
websockets==12.0
aiofiles==23.2.1
cryptography==42.0.7
email-validator==2.1.1
